backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.58.0
httpx==0.28.0
python-dotenv==1.0.1
